+++ main.py
@@ -173,6 +173,10 @@ class Main(KytosNApp):
 
         endpoint = find_endpoint(switch, port)
         if endpoint is None:
+            log.warning(f"Port {port} not found on switch {switch}")
+            return None
+        endpoint = endpoint['endpoint']
+        if endpoint is None:
             return {'out_port': port,
                     'entries': entries}